# Optimise sprintX functions in types.go