--- a/msg_helpers.go
+++ b/msg_helpers.go
@@ -6,7 +6,6 @@ import (
 	"encoding/binary"
 	"encoding/hex"
 	"net"
-	"strconv"
 	"strings"
 )
 
@@ -276,13 +275,7 @@ func unpackString(msg []byte, off int) (string, int, error) {
 			s.WriteByte('\\')
 			s.WriteByte(b)
 		case b < ' ' || b > '~': // unprintable
-			var buf [3]byte
-			bufs := strconv.AppendInt(buf[:0], int64(b), 10)
-			s.WriteByte('\\')
-			for i := len(bufs); i < 3; i++ {
-				s.WriteByte('0')
-			}
-			s.Write(bufs)
+			writeByte(&s, b)
 		default:
 			s.WriteByte(b)
 		}