--- a/types.go
+++ b/types.go
@@ -513,13 +513,10 @@ func appendByte(s []byte, b byte) []byte {
 	var buf [3]byte
 	bufs := strconv.AppendInt(buf[:0], int64(b), 10)
 	s = append(s, '\\')
-	for i := 0; i < 3-len(bufs); i++ {
+	for i := len(bufs); i < 3; i++ {
 		s = append(s, '0')
 	}
-	for _, r := range bufs {
-		s = append(s, r)
-	}
-	return s
+	return append(s, bufs...)
 }
 
 func nextByte(b []byte, offset int) (byte, int) {