--- a/types_test.go
+++ b/types_test.go
@@ -72,3 +72,21 @@ func TestSplitN(t *testing.T) {
 		t.Errorf("failure to split 510 char long string: %d", len(xs))
 	}
 }
+
+func TestSprintName(t *testing.T) {
+	got := sprintName("abc\\.def\007\"\127@\255\x05\xef\\")
+
+	if want := "abc\\.def\\007\\\"W\\@\\173\\005\\239"; got != want {
+		t.Errorf("expected %q, got %q", got, want)
+	}
+}
+
+func BenchmarkSprintName(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		got := sprintName("abc\\.def\007\"\127@\255\x05\xef\\")
+
+		if want := "abc\\.def\\007\\\"W\\@\\173\\005\\239"; got != want {
+			b.Fatalf("expected %q, got %q", got, want)
+		}
+	}
+}