This is #746 but for the `sprintX` functions in types.go. It follows the same principal.

There is a lot of overlap between these functions and I was hoping to merge them together, but it seemed like too much effort because they were all just a little too different from each other.

```
name               old time/op    new time/op    delta
SprintName-12         247ns ± 3%     218ns ± 3%  -11.54%  (p=0.000 n=9+10)
SprintTxtOctet-12     199ns ± 7%     198ns ±14%     ~     (p=0.780 n=10+10)
SprintTxt-12          384ns ± 6%     257ns ± 7%  -33.03%  (p=0.000 n=10+10)
UnpackString-12       101ns ± 6%     112ns ± 9%  +10.43%  (p=0.000 n=19+10)

name               old alloc/op   new alloc/op   delta
SprintName-12         80.0B ± 0%     48.0B ± 0%  -40.00%  (p=0.000 n=10+10)
SprintTxtOctet-12     80.0B ± 0%     80.0B ± 0%     ~     (all equal)
SprintTxt-12           168B ± 0%       80B ± 0%  -52.38%  (p=0.000 n=10+10)
UnpackString-12       48.0B ± 0%     48.0B ± 0%     ~     (all equal)

name               old allocs/op  new allocs/op  delta
SprintName-12          3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
SprintTxtOctet-12      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
SprintTxt-12           5.00 ± 0%      2.00 ± 0%  -60.00%  (p=0.000 n=10+10)
UnpackString-12        2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```