@miekg 2018-09-27
solving this merge also undos the later merge of #755